+++ src/pages/TeamSettings/FlowGroups.vue
@@ -84,6 +84,7 @@ export default {
     }
   },
   computed: {
+    ...mapGetters('license', ['hasPermission']),
     ...mapGetters('tenant', ['tenant']),
     ...mapGetters('api', ['isCloud']),
 
@@ -111,8 +112,11 @@ export default {
     isLoadingTable() {
       return this.loadingKey > 0
     },
-    isTenantAdmin() {
-      return this.tenant.role === 'TENANT_ADMIN'
+    permissionsCheck() {
+      return (
+        this.hasPermission('update', 'run') &&
+        this.hasPermission('delete', 'run')
+      )
     },
     items() {
       if (!(this.versionGroups && this.flows)) return []
@@ -246,7 +250,7 @@ export default {
     <template #title>Flow Groups</template>
 
     <template #subtitle>
-      <span v-if="isTenantAdmin">
+      <span v-if="permissionsCheck">
         View and manage your team's flows by
         <ExternalLink
           href="https://docs.prefect.io/cloud/concepts/flows.html#versioning"
@@ -404,7 +408,7 @@ export default {
           </template>
 
           <!-- FVG ACTIONS -->
-          <template v-if="isTenantAdmin" #item.action="{ item }">
+          <template v-if="permissionsCheck" #item.action="{ item }">
             <v-btn
               color="error"
               text