+++ Changelog.md
@@ -5,6 +5,7 @@
 ### Features and Improvements
 
 - Add agent screens - [#852](https://github.com/PrefectHQ/ui/pull/852)
+- Replaces everywhere we are using this.tenant.role with the hasPermission check - [#892](https://github.com/PrefectHQ/ui/pull/892)
 - Add the `mapped_children` route to the IAPI allow-list - [#891](https://github.com/PrefectHQ/ui/pull/891)
 - Add the netlify script to the csp - [#897](https://github.com/PrefectHQ/ui/pull/897)
 - Improve expired auth and id token handling - [#896](https://github.com/PrefectHQ/ui/pull/896)
@@ -14,6 +15,7 @@
 - Fix KV copy button issue - [#887](https://github.com/PrefectHQ/ui/pull/887)
 - Check explicitly for map index of 0 to avoid falsey in restart button [#890](https://github.com/PrefectHQ/ui/pull/890)
 - Remove setAgents from app.vue - [#898](https://github.com/PrefectHQ/ui/pull/898)
+- None
 
 ## 2021-06-02
 
@@ -33,7 +35,6 @@
 
 ### Bugfixes
 
-- Display task run name in Schematic - [#863](https://github.com/PrefectHQ/ui/pull/863)
 - Various bugfixes related to null tenants and team invitaitons - [#862](https://github.com/PrefectHQ/ui/pull/862)
 
 ## 2021-05-25